+++ lib/rsm/install/rails.rb
@@ -1,14 +1,7 @@
 module Rsm
   module Install
-    class Rails < Thor::Group
-      include Thor::Actions
-      include Actions
-
-      attr_reader :application_root, :worker_processes
-
-      argument :name
-
-      class_option :apps_root, :defualt => "/var/www", :aliases => "-r", :desc => "Rails apps root (default: /var/www)"
+    class Rails < Rsm::Base
+      attr_reader :worker_processes
 
       class_option :user, :aliases => "-u", :default => "git", :desc => "Owners's user"
       class_option :group, :aliases => "-g", :default => "git", :desc => "Owners's group"
@@ -19,17 +12,12 @@ module Rsm
 
       class_option :worker_processes, :type => :numeric, :default => 2, :aliases => "-w", :desc => "Worker processes for use in Unicorn"
 
-      def self.source_root
-        File.expand_path('../../../..', __FILE__)
-      end
-
       def set_destination_root
-        @application_root = "#{options[:apps_root]}/#{name}"
-        self.destination_root = @application_root
+        self.destination_root = application_root.to_s
       end
 
       def download
-        empty_directory application_root
+        empty_directory "."
         inside application_root do
           if options[:git]
             run "git clone #{options[:git]} ."
@@ -40,14 +28,13 @@ module Rsm
             fetch_temporary_archive(name, options[:tbz2], :bz2)
             unpack_compessed_archive(name, :bz2)
           else
-            say "No source URI specified. Use --git, --tgz or --tbz2 option with URI passed"
-            exit 1
+            say "No source URI specified - skip download"
           end
         end
       end
 
       def permissions
-        inside application_root do
+        inside "." do
           empty_directory "log"
           empty_directory "tmp"
           run "chown #{options[:user]}:#{options[:group]} -R ."
@@ -57,7 +44,7 @@ module Rsm
 
       def unicorn_config
         @worker_processes = options[:worker_processes]
-        template "templates/unicorn.rb.erb", "config/unicorn.rb"
+        template "unicorn.rb.erb", "config/unicorn.rb"
       end
     end
   end